+++ api/v1alpha1/terminatinggateway_types.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
+	"github.com/hashicorp/consul-k8s/namespaces"
 	capi "github.com/hashicorp/consul/api"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -186,6 +187,23 @@ func (in *TerminatingGateway) Validate(namespacesEnabled bool) error {
 			in.KubernetesName(), errs)
 	}
 	return nil
+}
+
+// DefaultNamespaceFields sets the namespace field on spec.services to their default values if namespaces are enabled.
+func (in *TerminatingGateway) DefaultNamespaceFields(consulNamespacesEnabled bool, destinationNamespace string, mirroring bool, prefix string) {
+	// If namespaces are enabled we want to set the namespace fields to their
+	// defaults. If namespaces are not enabled (i.e. OSS) we don't set the
+	// namespace fields because this would cause errors
+	// making API calls (because namespace fields can't be set in OSS).
+	if consulNamespacesEnabled {
+		// Default to the current namespace (i.e. the namespace of the config entry).
+		namespace := namespaces.ConsulNamespace(in.Namespace, consulNamespacesEnabled, destinationNamespace, mirroring, prefix)
+		for i, service := range in.Spec.Services {
+			if service.Namespace == "" {
+				in.Spec.Services[i].Namespace = namespace
+			}
+		}
+	}
 }
 
 func (in LinkedService) toConsul() capi.LinkedService {